Added gasket STL and scad file
The STL file is for a gasket to be printed in TPU (95A or softer) to be placed between an adapter to slot into the holes. It is not really necessary for these fans, but it looks pretty cool.
Designed for Arctic P14.
The scad file allows for the creation of gaskets for various sizes of fans: simply adjust the parameters.